functions/requirements.txt
firebase-admin>=6.2.0
firebase-functions>=0.1.0
pytest>=7.4.0
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.10.0
